libft

+++ libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: obalaban <obalaban@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/23 17:46:22 by obalaban          #+#    #+#             */
-/*   Updated: 2022/06/23 18:35:30 by obalaban         ###   ########.fr       */
+/*   Updated: 2022/06/29 14:37:38 by obalaban         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int	index;
+	size_t	index;
 
 	index = 0;
 	while (index < len)
